Embeddings/requirements.txt
Flask
gunicorn
sentence-transformers
Pillow